Incorporate Veronika's suggestions
